schema: Testing data gen functions.
The data generation functions are broken into
a separate file and an accompanying test file with
both some unit tests and some benchmarks were added.
